third_party/WebKit/LayoutTests/external/wpt/dom/events/relatedTarget.window-expected.txt

This is a testharness.js-based test.
PASS Reset if target pointed to a shadow tree
FAIL Reset if relatedTarget pointed to a shadow tree assert_equals: expected null but got object "[object XMLHttpRequest]"
PASS Reset if target pointed to a shadow tree pre-dispatch
PASS Reset if relatedTarget pointed to a shadow tree pre-dispatch
FAIL Reset targets on early return assert_false: expected false got true
FAIL Reset targets before activation behavior assert_equals: expected null but got Element node <input type="checkbox"></input>
Harness: the test ran to completion.
